[F] Fix long cannot be cast to int

+++ src/main/java/org/hydev/veracross/analyzer/api/nodes/veracross/NodeCourses.java
@@ -118,7 +118,7 @@ public class NodeCourses extends JsonApiNode<NodeCourses.Model>
             // Add course id
             if (!info.courseIds().contains("" + veraCourse.getId()))
             {
-                info.addCourseId((int) veraCourse.getId());
+                info.addCourseId(veraCourse.getId());
             }
 
             // Save and get info id
